requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fpdf2>=2.7.0
androguard>=4.1.0

